# openapi.json
{
  "components": {
    "responses": {
...
      "Error401": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            }
          }
        },
        "description": "Authentication is required or the supplied credential is invalid."
      },
      "Error403": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            }
          }
        },
        "description": "The authenticated principal is not permitted to perform this operation."
      },
      "Error404": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            }
          }
        },
        "description": "The requested tenant or resource was not found."
      },
...
      "Error422": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            }
          }
        },
        "description": "The request did not pass structural or semantic validation."
      },
...
    },
    "schemas": {
...
      "Page_RegistrationAdminSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RegistrationAdminSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "Page[RegistrationAdminSummary]",
        "type": "object"
      },
...
      "RegistrationAdminUpdate": {
        "properties": {
          "internal_notes": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Notes"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "SUBMITTED",
                  "CANCELLED"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "RegistrationAdminUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Domain-resolved, RLS-protected multi-tenant backend.",
    "title": "Kanoon Educational Institute API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/registrations": {
      "get": {
        "operationId": "list_registrations_api_v1_admin_registrations_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RegistrationAdminSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Registrations",
        "tags": [
          "admin-registrations"
        ]
      }
    },
...
    "/api/v1/admin/registrations/{registration_id}": {
      "get": {
        "operationId": "registration_detail_api_v1_admin_registrations__registration_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "registration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registration Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationAdminDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Registration Detail",
        "tags": [
          "admin-registrations"
        ]
      },
      "patch": {
        "operationId": "update_registration_api_v1_admin_registrations__registration_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "registration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registration Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationAdminUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationAdminDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Registration",
        "tags": [
          "admin-registrations"
        ]
      }
    },
...
    "/api/v1/public/registrations": {
      "post": {
        "operationId": "create_registration_api_v1_public_registrations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          }
        },
        "summary": "Create Registration",
        "tags": [
          "public-registrations"
        ]
      }
    },
    "/api/v1/public/registrations/{registration_id}": {
      "get": {
        "operationId": "get_registration_api_v1_public_registrations__registration_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "registration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          }
        },
        "summary": "Get Registration",
        "tags": [
          "public-registrations"
        ]
      },
      "patch": {
        "operationId": "patch_registration_api_v1_public_registrations__registration_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "registration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          }
        },
        "summary": "Patch Registration",
        "tags": [
          "public-registrations"
        ]
      }
    },
    "/api/v1/public/registrations/{registration_id}/contacts": {
      "put": {
        "operationId": "put_contacts_api_v1_public_registrations__registration_id__contacts_put",
        "parameters": [
          {
            "in": "path",
            "name": "registration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationContactsPut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RegistrationContactResponse"
                  },
                  "title": "Response Put Contacts Api V1 Public Registrations  Registration Id  Contacts Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          }
        },
        "summary": "Put Contacts",
        "tags": [
          "public-registrations"
        ]
      }
    },
    "/api/v1/public/registrations/{registration_id}/otp/send": {
      "post": {
        "operationId": "send_otp_api_v1_public_registrations__registration_id__otp_send_post",
        "parameters": [
          {
            "in": "path",
            "name": "registration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OTPSendResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "summary": "Send Otp",
        "tags": [
          "public-registrations"
        ]
      }
    },
    "/api/v1/public/registrations/{registration_id}/otp/verify": {
      "post": {
        "operationId": "verify_otp_api_v1_public_registrations__registration_id__otp_verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "registration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OTPVerifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          }
        },
        "summary": "Verify Otp",
        "tags": [
          "public-registrations"
        ]
      }
    },
...
    "/api/v1/public/registrations/{registration_id}/profile-image/upload": {
      "post": {
        "operationId": "initiate_profile_image_upload_api_v1_public_registrations__registration_id__profile_image_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "registration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadInitiateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadInitiateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          }
        },
        "summary": "Initiate Profile Image Upload",
        "tags": [
          "public-registrations"
        ]
      }
    },
    "/api/v1/public/registrations/{registration_id}/profile-image/{media_id}/complete": {
      "post": {
        "operationId": "complete_profile_image_upload_api_v1_public_registrations__registration_id__profile_image__media_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "registration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registration Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "media_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Media Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          }
        },
        "summary": "Complete Profile Image Upload",
        "tags": [
          "public-registrations"
        ]
      }
    },
...
  }
}